Create README.md
Douglab is a repository to share ansible playbooks, shell scripts and python codes to public, also sharing automation codes to the job interview and my own linux environment.
Welcome to our git repository douglab-repo1. As a guest user, please be aware that this repository is provided solely for educational and study purposes. We cannot guarantee the accuracy, completeness, or suitability of any of the information or tools provided on this server.
Please note that any code or scripts executed on this server are done so at your own risk. We strongly recommend that you ensure that any code you run or scripts you execute have been tested in a non-production environment before running them on any live systems.
By using this public git repository nominated as douglab-repo1, you acknowledge that you accept all responsibility and risks associated with your use of this any server. Any damages or losses incurred as a result of using this server are the sole responsibility of the user.
Thank you for using our Jenkins server for your educational and study purposes.